Even Liber Abbaci

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -7,8 +7,8 @@
 int main(void)
 {
 	int sum;
-	long int temp;
-	long int x = 1, y = 2;
+	int temp;
+	int x = 1, y = 2;
 
 	sum = 2;
 	while (temp <= 4000000)
